add validation

+++ controllers/courseController.js
@@ -185,8 +185,11 @@ exports.getDetailCourse = async (req, res) => {
 
 exports.softDeleteCourse = async (req, res) => {
   const { id } = req.params;
-  const {deleted_by} = req.body
+  const { deleted_by } = req.body
   try {
+    if (!deleted_by) {
+      return res.status(404).json({ success: false, message: 'User required!.' });
+    }
     await deleteCourse(deleted_by, id)
     res.status(200).json({
       success: true,
